feat: simplified build tools
Frontend restructured, is nothing more than a tailwind builder now.
Rely on the serverside templates for this one.
I don't see much need for other frameworks other than HTMX partials,
which are not part of a frontend build, either.

Slight edits to posts templates while I feel things out.

Added vscode setting for label names in editors.
With so many files with the same name,
it helps to see which directory they're in.

Brewfile updated for tailwindcss CLI.
Justfile bootstrap now installs cargo-watch, as well,
which is depended on for `just up`.

Still would need a second window running the Tailwind dev build,
but it's also possible an overmind procfile could be introduced?
Or go into Tilt, instead.