feat: collection of updates
Markdown filter implemented into the Tera engine.
Honestly I think this needs to be externalized due to the annoying difficulty of the thing,
we'll see where that lands.

Template macros implemented for some button and form stuff.
Though, I am moving in the direction of using a UI lib of some kind to replace it.